Changed directory structure -> Pages scattered in various places are now stored in pages Creating and testing control directories -> Added a Contents directory -> Started creating a pagemove control
