+++ OnePunchMan/Contents/PageMove.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace OnePunchMan.Contents;
+
+public partial class PageMove : ContentView
+{
+	public PageMove()
+	{
+		InitializeComponent();
+	}
+}
