--- OnePunchMan/MainPage.xaml.cs
@@ -1,31 +1,0 @@
-﻿namespace OnePunchMan;
-
-public partial class MainPage : ContentPage
-{
-	int count = 0;
-
-	public MainPage()
-	{
-		InitializeComponent();
-	}
-
-	private void OnCounterClicked(object sender, EventArgs e)
-	{
-		count++;
-
-		if (count == 1)
-			CounterBtn.Text = $"Clicked {count} time";
-		else
-			CounterBtn.Text = $"Clicked {count} times";
-
-		SemanticScreenReader.Announce(CounterBtn.Text);
-	}
-
-    private async void OnClicked(object sender, EventArgs e)
-    {
-        // サブページへ移動
-        await Shell.Current.GoToAsync("//about");
-    }
-}
-
-
