+++ OnePunchMan/Pages/Activity_Counter.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace OnePunchMan.Pages;
+
+public partial class Activity_Counter : ContentPage
+{
+	public Activity_Counter()
+	{
+		InitializeComponent();
+	}
+}
