+++ OnePunchMan/Pages/Activity_Running.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace OnePunchMan.Pages;
+
+public partial class Activity_Running : ContentPage
+{
+	public Activity_Running()
+	{
+		InitializeComponent();
+	}
+}
