+++ OnePunchMan/Pages/Home.xaml
@@ -1,0 +1,19 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="OnePunchMan.Pages.Home"
+             Title="Home">
+    <VerticalStackLayout>
+        <Label 
+            Text="ワンパンマンチャレンジャー"
+            VerticalOptions="Center" 
+            HorizontalOptions="Center" />
+
+        <Button
+                x:Name="CounterBtn"
+                Text="Go to About"
+                SemanticProperties.Hint="Counts the number of times you click"
+                Clicked="moveToAbout"
+                HorizontalOptions="Center" />
+    </VerticalStackLayout>
+</ContentPage>
