+++ OnePunchMan/Pages/Settings.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace OnePunchMan.Pages;
+
+public partial class Settings : ContentPage
+{
+	public Settings()
+	{
+		InitializeComponent();
+	}
+}
